requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
sqlmodel
sqlalchemy-utils
requests
python-dotenv
psycopg2-binary
email-validator
sentry-sdk
Werkzeug
cryptography
itsdangerous
binance-connector
aiocache
aioredis


